Perl sucks and here's why. The latest version of perl supplies MakeMaker version 6.57, but some of the support packages require version 6.62.
So we have to build newer version of MakeMaker. But MakeMaker is required to
build perl in the first place. Can you drop in the 6.62 tree into the
perl tree and update the source? Of course not.

Instead, you have create an alternate package. But now the perl and makeMaker rpms collide.  basically, no win.  Having software rely on versions beyond the latest release is brain dead. Welcome to bio and perl code.