Dependencyless cpan rpm packages - release version configurable
Now the packages generated by cpan should really be dependecyless
we have to specify the nodeps when building the rpm

Added some code to configure dinamically the release version of the
rpm packages. If the RocksReleaseVersion environment variable is set
the RPM will use that variable as a release version.